feat: :sparkles: 3. Asignación de Asientos
feat: :construction: Successful test to validate the reservation of a seat for a screening

It was validated that the seat exists, that the projection exists and that that seat is available. Additionally, when making the reservation, the payment method is added and the seat is updated to reserved, just as in the purchase collection the data is updated.

feat: :construction: Successful test to cancel a seat reservation for a screening

It was taken into account that the purchase id existed in the purchase collection and the seat id to proceed to cancel the reservation, immediately the purchase and ticket data were updated and the status of that seat became available

feat: :white_check_mark: test of case 3 passed successfully

Several aspects were taken into account to validate a seat in an available state to be reserved and also to be able to cancel that reservation and return that seat to an available state.